Sort: Use cub::WarpMergeSort for small sorts (32 < n <= 128)
We currently use `bitonicSortKVInplace` for sorts of size `n <= 32`
but use `radixSortKVInplace` for `32 < n <= 4096`. Bitonic sort is
also unstable, which forces stable sorts fall back to which is up to
4x slower in this small regime.

This PR adds a new kernel `warpMergeSortKVInplace` using
`cub::WarpMergeSort` to implement sorts with `32 < n <= 128` and all
stable sorts with `n < 128`. This results in up to a 2x speedup for
unstable sorts and up to 15x for stable sorts, depending on the input
geometry.

In terms of number of kernels, this replaces radix-sorts of size 32
and 128 WarpMergeSort of size 128 both ascending and descending so is
a wash.

ghstack-source-id: 63a886cc87299778d4d2bc2434d9069097f2848a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/96223